Some fixes for terms and added some limits to functionality if user didnt accept the terms

+++ tgbot/handlers/search_music.py
@@ -5,12 +5,18 @@ from aiogram.types import InlineKeyboardMarkup, InlineKeyboardButton, ContentTyp
 from aiogram.utils.exceptions import MessageIsTooLong
 from youtubesearchpython import SearchVideos
 
+from tgbot.config import Config
 from tgbot.handlers.user import run_blocking_io
 from tgbot.keyboards.callback_datas import video_callback
+from tgbot.keyboards.inline import accept_terms_keyboard
 from tgbot.models.db_utils import Database
 
 
-async def search_music_func(mes: types.Message, db: Database):
+async def search_music_func(mes: types.Message, db: Database, config: Config):
+    is_accepted = await db.check_user_terms(mes.from_user.id)
+    if is_accepted is False:
+        await mes.answer(config.terms.cond_terms_text, reply_markup=accept_terms_keyboard)
+        return
     try:
         await mes.delete()
     except Exception:
